Added the extended benchmark (`exbench.lua`) that currently has MD5 and SHA256 checksum tests.
This script benchmarks MD5 and SHA256 checksum computation for files. It allows each benchmark to be skipped if an error occurs, otherwise it runs multiple iterations of the file checksum (using the entry file as a sample) to measure performance. `exbench.lua` is called by `bench.lua` to keep basic benchmarks separated from more complex and dependency-driven ones.